requirements.txt
python-dotenv==1.0.0
Flask==2.3.3
requests==2.31.0
Pillow==10.0.0
twilio==9.5.1